Started Drag&Drop System, changed ScrollBars + various smaller changes
Work on a Drag&Drop system was started with the PDnDSupport and
PDnDTransfer interfaces. A first implementation for the PList was also
started.
PListModel was enhanced with methods to manipulate the content. This
will be needed for D&D and Copy, Cut, Paste support.
ReRendering system in Swing implementation was changed to repaint only
what is necessary.
PTable got broken, needs to be fixed in a future update.
PPicture was enhanced with a flag to set whether the picture should be
stretched to fit size when rendering.
PList got a new method to query element index based on coordinates.