Continued work on Drag & Drop system
Created a Drag & Drop manager class that can be added to PRoots to
support dragging and dropping of PComponents across a PGlassPanel.
On the side of the PRoot neither the PGlassPanel nor the DnDManager have
been added yet, this has to come in the future.
Also: Some smaller fixes / changes / updates to a few classes.