Improvements in isMouseOver
The isMouseOver and isMouseOverThisOrChildren is now using the
getComponentAtMouse method in the PMouse instead of calculating the
component at mouse position.
This could improve performance if the PMouse instance caches the
component it is currently placed on top of.
The Swing implementation makes use of such a caching technique. The
PCompUtil class can still be used by the PMouse implementation if
caching is not wanted.