Added support for PLayoutDesigns
Layouts can now have designs that might change certain attributes of the
layout. For example the insets or the gap might be defined by the
design.
For this purposes the PLayoutDesign interface was added, the
PDesignSheet, PDesignFactory and PReadOnlyLayout interfaces were
changed.

Several of the already existing layouts now use this new system for
their attributes but the work is still in progress.