Changes at DnDSupport + DnD for PList
DnDSupport now has an additional method to retrieve the active
DnDTransfer. The documentation has also been updated with additional
exceptions.
PMouse has been updated to include a constant for the Drag-and-Drop
mouse button.
The PDropComponent now has additional methods needed for DnD.

The PList can now be used with DnD.

The DefaultPDnDSupport and the Swing implementation of PMouse has been
updated to reflect the above changes.