Changed DnD system
The PDnDTransfer now returns a List of Index-and-Content tuples instead
of an Object when querying the transfer data.
This is done to preserve structural information about the data when a
transfer is performed. A PDnDSupport of a Tree can now export its data
with additional information about the tree structure of the data. The
receiving DnDSupport can decide to use this additional information or
not.