Added ObserverList interface
A new interface for ObserverList's was added with two implementations.
These lists should replace the current notion of using
CopyOnWriteArrayList's as observer lists throughout the library.
Instead, the createDefaultObserverList method in the PCompUtil class
should be used to create an instance of ObserverList.

Changing the observer list in all core classes and components will be
happening over the next few commits until all use the new system.