Added PModelFactory class
The PModelFactory should be incorporated into all core components that
use models.
The model factory will provide default implementations for all core
models via static methods. The static ModelFactory instance that will be
used can be set via a public static method within this class.
This API is supposed to be used for a simpler way of overriding the
models of all core components to custom implementations.

This class is still work-in-progress and not yet being used by most core
components.