Remade PTable + Improvements to Layouting system + Various small changes
- PCellComponent now remembers the PModel and PModelIndex set.
- PCellComponentWrapper and DefaultPCellComponent implement these
changes.
- Added method to PComponent to request focus without taking it away
from own descendants. (PList and PTable use this now)
- Added method to PComponent to get constraint with which component is
registered at parent.
- Made methods of PModelObs to default methods.
- Changed the way invalidation of layout data is handled to reduce
unnecessary method calls. Layouts now have to invalidate themselves when
a child is removed. When the preferred size of a child changes the
onChildPrefSizeChanged method of the layout is called and the layout can
decide whether to invalidate itself or not.
- Many layouts reworked to reduce number of invalidation of layout data.
- PSplitLayout now has a get-method to get the Divider.
- Remade a new version of PTable, PTableModel and PTableLayout. Still
WIP.