Bug Fixes + new Layout System + Small improvements
- Fixed bugs which appeared in the PDropDown and PComboBox as a result
of previous changes. The PDropDown is now triggered on a model change
from true to false instead of a click detected by a MouseObs.
- Improved the constructors for the PSpinnerModelList,
DefaultPSliderModel and DefaultPListModel
- Added convenient methods to PList to query and manipulate the
selection directly without having to go over the PSelection
- Small bug fix in the PCellComponentWrapper which did incorrect
rendering in some cases
- Added an isEmpty() method to PSize to check for either width or height
being zero and added a set(PSize) method to MutablePSize as a
copy-setter
- Added convenience methods to AbstractPContainer which were intended to
be there but previously missing
- Added extension methods to AbstractPComponent to make it easier to
subclass. The AbstractPComponent now has methods which are called when
the root changes, the parent changes or the component is laid out. The
default toString() method has also been improved.
- Small fix to the AbstractPBorder whichs behavior has been changed
unintentionally in a previous update. A border will now grow its
internal contents by default.
- Further improvements to the new layout system. The PLayout,
AbstractPLayout and AbstactPLayoutOwner classes have been modified to
reflect these changes. Its now easier to cache information within a
layout to reduce the number of calculations being done.
- Changed all layouts to reflect the changes to the layouting system.
- Improved the existing Swing tests and updated them to reflect the most
recent changes