Small improvements + Bug fixes + Continued work on InnerPRoot
- Fixed a small bug in PComponent method isMouseOverThisOrChild() where
a possible null return value was not considered
- Moved the invalidate() method to PReadOnlyLayout to make it more
accessible. This change might not be final.
- Since invalidate is now public for every layout a bug in the
reLayOutTheEntireGui() method of PRoot could be fixed now. The same is
true for the reLayOut() method of AbstractPRoot.
- The platform independent part of the working render method of the
JCompPRoot has now been moved to the AbstractPRoot so it can be reused
by other root implementations
- Changed all fields of AbstractPRoot from private accessibility to
protected
- Improved the InnerPRoot and DummyPRoot (now called DelegatePRoot)
- Added a copy-setter method to the MutablePBounds just like there is
one for the MutablePSize.
- Small refactoring for PCompUtil and SwingPMouse.