Big Update
- Reworked the PKeyInput system to use generics for better ease of use.
- Heavily refactored the Drag&Drop system. A component can now more
easily be used as a Drag&Drop indicator. Also improved the PDnDSupport
and PDnDManager classes.
- Reworked the PTable, PTree and PList with all accompanying classes
(big changes).
- Reworked the PTimer system. Now the detla time (in milliseconds) since
the last update is passed as an argument to the PTimerCallback. The
PRoot now also has a method to get the detla time.
- Small performance improvement in PSplitPanel
- Fixed a small bug with PSpinner parts not being disabled properly when
the PSpinner was disabled.
- Added another utility method to PCompUtil which will return the first
ancestor component of a given component which is of a given type.
- Updated all Swing tests and the tutorial classes.