Update & Bugfixes
- Fixed a bug introduced by the previous update where the setModel
method for various components was called too early within their
constructors.
- Added a clickCount to the onButtonTriggered method.
- Changed the delay of PTimer to be a double. Since the time is measured
as a double it only makes sense to make the delay a double too.
- Fixed the default value for the PCaretRenderTimer focusToggleTimer. It
was way too short before. Also improved the focusToggleTimer
implementation as it was unnecessarily complicated before.
- Added the method getRenderPositionForIndex to PTextComponent which
returns the bounds for the char at the given index. Still WIP and will
probably be changed in future updates to take a start and end index for
an entire String instead.
- Improved default preferred size calculations for PTextArea and
PTextField.
- Added method onScrollRequest to PComponentObs which will be called if
scrolling within the ancestor tree of the component is in order. Also
added methods to AbstractPComponent to issue scroll requests.
- Added default methods to PMouse to calculate the offset of the current
mouse position to a given component.
- Fixed a critical bug in AbstractPComponent which would cause
components to not be correctly laid out if added to the GUI tree in an
unlucky order.
- Improved rendering logic of AbstractPRoot. Components will no longer
render on top of their borders by default. Also eliminated some more
unnecessary rendering calls.
- Added an absolute set-method to MutablePBounds.