Big Update
- Reworked the layouting subsystem. Every component can now declare a
PLayoutPreference. Layouts are free to honor this preference or ignore
it. Layouting data will now be saved in a PLayoutData object.
PLayoutData replaces and improves the PCompInfo class previously used by
all layouts. AbstractPComponents now cache a reference to their
PLayoutData directly for faster access of this important data.
- Added a new subsystem to represent abstract component actions like
copy & paste, delete, moving cursors, etc. The system has ties to the
menu subsystem to support calling these actions from popups and menus.
The system is still a WIP and some changes need to be made in future
updates.
- Reworked the Popup and Menu subsystems. Popups already work very well,
menus are still a WIP.
- Created a top-level interface for many simple models like the
TextModel, CheckBoxModel or PictureModel. All of these models now
implement the PSingleValueModel. For this reasons observers have also
been changed. Still a WIP.
- Many small bug fixes, improvements and refactorings including:
* Made PListIndex Comparable
* Added many new useful default methods to PSelectionComponent
* Added overloaded constructor to PTableCellIndex
* Fixed a bug in the constructor of PTree
* Fixed a cache related bug in DefaultPTextModel
* Added new, convenient constructors to PButton and PCheckBoxTuple
* Added a clear() method to PRadioButtonGroup
* Removed PPassWordField and replaced it with a PPasswordTextModel. A
new component did not make sense for this functionality since it can
easily be done by a text model alone.
* The children List of layouts and components has now become an Iterable
* A new alignment PREFERRED_OR_CENTER has been added for the new layout
preferences.
* Many small improvements have been made to various layouts as part of
the layout subsystem overhaul.
* Fixed a bug in the layout calculations of PWrapLayout
* Added a few more convenient methods to PComponent to iterate the GUI
tree. Existing methods that iterate ancestors have been changed to
include the calling object.
* Key constants in PKeyboard have been split into ActualKey and
VirtualKey which both implement the interface Key. New key constants can
be defined by the user by implementing the interface as needed.
* More extension methods have been added to AbstractPComponent to make
subclassing easier. onAddedToUi is called when a component is added to a
UI without having been part of a UI before. onRemovedFromUI is called
when a component has been removed from a UI without having been moved to
a new one. They are special cases of onRootChanged which is still called
in both of these cases as well as others.
* An iterator implementation has been added to AbstractPListModel
* A new API for error handling has been introduced but has not yet been
incorporated into the code base. The new API will be added incrementally
in future updates.