Continued work on style subsystem + small fixes and improvements
- Added new Annotations @CallSuper and @TemplateMethod which will be
used to document source code.
- Removed the PStyleLayout and all of its uses. A replacement is not
planned as of now but might be added at some point in the future.
- Renamed method getStyleFromSheet() of PStyleable to
getInheritedStyle(). The same goes for the respective setter.
- Created an AbstractPStyleable class which has become a super class of
both AbstractPComponent and AbstractPBorder.
- Added a common super interface PStyle for both PStyleComponent and
PStyleBorder
- Renamed onSizeChanged() in PStyleObs to onSizeChangedEvent() to keep
in line with the naming convention
- Added two abstract implementations of PStyle called MutablePStyle and
ImmutablePStyle
- Fixed a bug in the BufferedObsList
- Created new BufferedWeakRefObsList which is currently not usable since
it does not work with many classes in the core library.